+++ src/app/RootLayout.tsx
@@ -3,34 +3,34 @@ import {
   Route,
   BrowserRouter as Router
 } from "react-router-dom";
-import { DragDropContext } from 'react-dnd';
-import HTML5Backend from 'react-dnd-html5-backend';
 import {
-  Timeline,
-  ManageSheriff
-} from './pages'
+  DragDropContext
+} from 'react-dnd';
+import HTML5Backend from 'react-dnd-html5-backend';
 import Footer from './components/Footer';
 import Navigation from './components/Navigation'
 import './assets/styles/Glyphicons.css'
 import './index.css';
+import Timeline from './pages/Timeline';
+import ManageSheriffs from './pages/ManageSheriffs';
 
 class Layout extends React.PureComponent {
   render() {
     return (
-        <Router>
-          <div className="App">
-            <div className="headerArea">
-              <Navigation />
-            </div>
-            <div className="mainArea">
-              <Route exact path='/' component={Timeline} />
-              <Route path='/sheriffs/manage' component={ManageSheriff} />
-            </div>
-            <div className='footerArea'>
-              <Footer />
-            </div>
+      <Router>
+        <div className="App">
+          <div className="headerArea">
+            <Navigation />
           </div>
-        </Router>
+          <div className="mainArea">
+            <Route exact path='/' component={Timeline} />
+            <Route path='/sheriffs/manage' component={ManageSheriffs} />
+          </div>
+          <div className='footerArea'>
+            <Footer />
+          </div>
+        </div>
+      </Router>
     );
   }
 }
