--- src/app/components/AssignmentTimeline/index.ts
@@ -1,1 +1,0 @@
-export { default as AssignmentTimeline, AssignmentTimelineProps, UNASSIGNED_GROUP, UNASSIGNED_ID } from './AssignmentTimeline'
