--- src/app/components/Form/index.ts
@@ -1,9 +1,0 @@
-export { default as TextFormField } from './BootstrapTextField';
-export { default as RequiredTrainingChecklist } from './RequiredTrainingChecklist';
-export { default as AssignmentTypeSelector } from './AssignmentTypeSelector';
-export {default as SubmitButton} from './SubmitButton';
-export { default as DateField } from './DateField';
-export { default as DateTimeField } from './DateTimeField';
-export { default as NumberOfSheriffsSelector} from './NumberOfSheriffsSelector';
-export { default as CheckboxField } from './BootstrapCheckboxField';
-export { default as TextArea } from './BootstrapTextArea';
