+++ src/app/components/FormElements/CourthouseSelector.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { default as FormFieldWrapper, FormFieldWrapperProps } from './Form/FormFieldWrapper';
+import { default as FormFieldWrapper, FormFieldWrapperProps } from './FormFieldWrapper';
 import { FormControl } from 'react-bootstrap';
