--- src/app/components/Timeline/index.ts
@@ -1,3 +1,0 @@
-export { default as Timeline, TimelineProps } from './Timeline';
-export {default as TimelineExtension} from './TimelineExtension'
-
