--- src/app/modules/assignments/index.ts
@@ -1,7 +1,0 @@
-export {default as reducer,AssignmentState as State} from './reducer';
-
-import * as _actions from './actions';
-export const actions = _actions;
-
-import * as _selectors from './selectors'
-export const selectors = _selectors;
