--- src/app/modules/sheriffs/index.ts
@@ -1,8 +1,0 @@
-export {default as reducer,SheriffState as State} from './reducer';
-
-import * as _actions from './actions';
-export const actions = _actions;
-
-import * as _selectors from './selectors';
-export const selectors = _selectors;
-
