--- src/app/modules/timeline/index.ts
@@ -1,5 +1,0 @@
-export { default as reducer, TimelineState as State } from './reducer';
-
-import * as _actions from './actions';
-export const actions = _actions;
-
