--- src/app/pages/index.ts
@@ -1,2 +1,0 @@
-export {default as Timeline} from './Timeline'
-export {default as ManageSheriff} from './ManageSheriffs';
